--- ghapd.config.json
@@ -1,4 +1,0 @@
-{
-    "todo": "implement this file",
-    "why nothing": "currently this file is needed to exist for ghapd to run but it is not queried in any way"
-}
