+++ pencil_pusher/configs/sources.py
@@ -1,7 +1,7 @@
-from ghapd.configs.manager import ConfigManager
 from typing import Any
-from ghapd.configs.base import ConfigParamBase
-from ghapd.configs import config_manager
+from pencil_pusher.configs.manager import ConfigManager
+from pencil_pusher.configs.base import ConfigParamBase
+from pencil_pusher.configs import config_manager
 
 
 @config_manager.register("sources")
