+++ pencil_pusher/documenter.py
@@ -1,4 +1,4 @@
-from ghapd.process_util import ProcessUtil
+from pencil_pusher.process_util import ProcessUtil
 import logging
