+++ tests/test_documenter.py
@@ -1,4 +1,4 @@
-from ghapd.documenter import Documenter
+from pencil_pusher.documenter import Documenter
 import os
 
 
@@ -7,6 +7,6 @@ def test_gen_local():
     Tests generating a doc with a
     local file
     """
-    Documenter.generate("ghapd.documenter", "test")
+    Documenter.generate("pencil_pusher.documenter", "test")
     assert os.path.isfile("./test.md")
     os.remove("./test.md")
